Issue #11163: Enforced filesize PositionSingleline//
