+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/abstractjavadoc/InputAbstractJavadocPositionWithSinglelineCommentsOne.java
@@ -1,0 +1,108 @@
+/*
+com.puppycrawl.tools.checkstyle.checks.javadoc.AbstractJavadocCheckTest$JavadocCatchCheck
+
+*/
+
+package com.puppycrawl.tools.checkstyle.checks.javadoc.abstractjavadoc;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/////////////
+// CLASSES //
+/////////////
+
+/**Javadoc*/ /*noise*/ //noise
+class/**nope*/ InputAbstractJavadocPositionWithSinglelineCommentsOne/**nope*/{
+    /**Javadoc*/ //noise
+    protected/**nope*/ class/**nope*/ N/**nope*/{/**nope*/}
+
+    /**Javadoc*/ //noise
+    private/**nope*/ static/**nope*/ class/**nope*/ X/**nope*/{/**nope*/}
+
+    /**Javadoc*/ //noise
+    @Component2/**nope*/ class/**nope*/ Z/**nope*/{/**nope*/}
+
+    /**Javadoc*/ //noise
+    @Component2/**nope*/ private/**nope*/ class/**nope*/ Y/**nope*/{/**nope*/}
+
+    /**Javadoc*/ //noise
+    private/**nope*/ @Component2/**nope*/ class/**nope*/ M/**nope*/{/**nope*/}
+}
+
+//////////////////
+// CONSTRUCTORS //
+//////////////////
+
+/**Javadoc*/ //noise
+class/**nope*/ VV/**nope*/{
+    /**Javadoc*/ //noise
+    VV/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+/**Javadoc*/ //noise
+class/**nope*/ NN/**nope*/{
+    /**Javadoc*/ //noise
+    private/**nope1*/ NN/**nope2*/(/**nope3*/)/**nope4*/{/**nope5*/}/**nope6*/
+}
+
+class/**nope*/ ZZ/**nope*/{
+    /**Javadoc*/ //noise
+    @Component2/**nope*/ ZZ/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+class/**nope*/ YY/**nope*/{
+    /**Javadoc*/ //noise
+    @Component2/**nope*/ private/**nope*/ YY/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+class/**nope*/ MM/**nope*/{
+    /**Javadoc*/ //noise
+    private/**nope*/ @Component2/**nope*/ MM/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+/////////////
+// METHODS //
+/////////////
+
+class VVV {
+    /**Javadoc*/ //noise
+    void/**nope*/ a/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+class NNN {
+    /**Javadoc*/ //noise
+    private/**nope*/ void/**nope*/ a/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+class XXX {
+    /**Javadoc*/ //noise
+    static/**nope*/ private/**nope*/ void/**nope*/ a/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+class ZZZ {
+    /**Javadoc*/ //noise
+    @Component2/**nope*/ void/**nope*/ a/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+class YYY {
+    /**Javadoc*/ //noise
+    @Component2/**nope*/ private/**nope*/ void/**nope*/ a/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+class MMM {
+    /**Javadoc*/ //noise
+    static/**nope*/ @Component2/**nope*/ private/**nope*/ void/**nope*/ a/**nope*/(/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+class LKJ {
+    /**Javadoc*/ //noise
+    void/**nope*/ a/**nope*/(@Component2/**nope*/int/**nope*/ a/**nope*/)/**nope*/{/**nope*/}/**nope*/
+}
+
+class LKM {
+    /**Javadoc*/ //noise
+    java.lang.String/**nope*/ a/**nope*/(/**nope*/)/**nope*/{/**nope*/return null;/**nope*/}/**nope*/
+}
